+++ src/components/Cell.js
@@ -9,12 +9,7 @@ import {observer} from 'mobx-react/native';
 export default class Cell extends React.Component {
     render() {
         const cell = (
-            <View
-                style={[
-                    {flexDirection: 'row', alignItems: 'center', padding: 15 * k},
-                    this.props.style,
-                ]}
-            >
+            <View style={[{flexDirection: 'row', alignItems: 'center', padding: 15 * k}, this.props.style]}>
                 {this.props.image &&
                     <View
                         style={[
@@ -45,9 +40,7 @@ export default class Cell extends React.Component {
                                     flex: 1,
                                     fontFamily: 'Roboto-Regular',
                                     fontSize: 15,
-                                    color: location.isDay
-                                        ? navBarTextColorDay
-                                        : navBarTextColorNight,
+                                    color: location.isDay ? navBarTextColorDay : navBarTextColorNight,
                                 },
                                 this.props.textStyle,
                             ]}
@@ -57,10 +50,7 @@ export default class Cell extends React.Component {
                     {typeof this.props.children !== 'string' && this.props.children}
                 </View>
                 {this.props.onRemove &&
-                    <TouchableOpacity
-                        style={{justifyContent: 'center', alignItems: 'center'}}
-                        onPress={this.props.onRemove}
-                    >
+                    <TouchableOpacity style={{justifyContent: 'center', alignItems: 'center'}} onPress={this.props.onRemove}>
                         <Image source={require('../../images/iconCloseSmall.png')} />
                     </TouchableOpacity>}