+++ src/components/EventBotMetabar.js
@@ -10,7 +10,7 @@ import statem from '../../gen/state';
 import autobind from 'autobind-decorator';
 
 type Props = {
-    bot: Bot,
+    bot: Bot
 };
 
 const styles = StyleSheet.create({
@@ -32,23 +32,20 @@ export default class extends React.Component {
         const bot = this.props.bot;
         const distance = location.location
             ? location.distanceToString(
-                location.distance(
-                    location.location.latitude,
-                    location.location.longitude,
-                    bot.location.latitude,
-                    bot.location.longitude
-                )
-            )
+                  location.distance(location.location.latitude, location.location.longitude, bot.location.latitude, bot.location.longitude)
+              )
             : null;
-        return (<View style={{flexDirection: 'row', height: 40 * k, alignItems: 'center'}}>
-            <View style={{padding: 10 * k}}><Image source={require('../../images/iconSubSmall.png')} /></View>
-            <View style={{paddingRight: 10 * k}}><Text style={styles.text}>{bot.followersSize}</Text></View>
-            <Separator />
-            <View style={{padding: 10 * k}}><Image source={require('../../images/iconPhotosHs.png')} /></View>
-            <View style={{paddingRight: 10 * k}}><Text style={styles.text}>{bot.imagesCount}</Text></View>
-            <View style={{flex: 1}} />
-            <View style={{padding: 10 * k}}><Image source={require('../../images/iconBotLocation2.png')} /></View>
-            <View style={{paddingRight: 15 * k}}><Text style={styles.text}>{distance}</Text></View>
-        </View>);
+        return (
+            <View style={{flexDirection: 'row', height: 40 * k, alignItems: 'center'}}>
+                <View style={{padding: 10 * k}}><Image source={require('../../images/iconSubSmall.png')} /></View>
+                <View style={{paddingRight: 10 * k}}><Text style={styles.text}>{bot.followersSize}</Text></View>
+                <Separator />
+                <View style={{padding: 10 * k}}><Image source={require('../../images/iconPhotosHs.png')} /></View>
+                <View style={{paddingRight: 10 * k}}><Text style={styles.text}>{bot.imagesCount}</Text></View>
+                <View style={{flex: 1}} />
+                <View style={{padding: 10 * k}}><Image source={require('../../images/iconBotLocation2.png')} /></View>
+                <View style={{paddingRight: 15 * k}}><Text style={styles.text}>{distance}</Text></View>
+            </View>
+        );
     }
 }