+++ src/components/Screen.js
@@ -9,12 +9,7 @@ import {observer} from 'mobx-react/native';
 export default class Screen extends Component {
     render() {
         return (
-            <View
-                style={[
-                    {position: 'absolute', top: 0, left: 0, right: 0, bottom: 0},
-                    this.props.style,
-                ]}
-            >
+            <View style={[{position: 'absolute', top: 0, left: 0, right: 0, bottom: 0}, this.props.style]}>
                 <BackgroundGradient isDay={location.isDay} />
                 <View style={[{flex: 1}]}>
                     {this.props.children}