+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Suggested subheadings: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Se
 
 If there are not many items, just list them sequentially. 
 
+* Source code formatting changes
 * Fix: Can't delete bots (#645)
 
 # 1.29.1 - 2017 May 4